A solution for model retrieval after program ends
Since till now, one has to use channels to communicate data from within
the bubbletea program to after it finished, a panic will block the hole
program, since the receiving channel block indefinitely.

This change would overcome the problem since communication is possible
**through** the program and the Model type of the user.

The use of a pointer to a empty interface (Model) is IMHO justified
through the setting of the model with the update cycle, which is a
performance bottle neck and if the user did not use a pointer to the
own struct, a pointer to it will be faster.